Fix a few bugs
This implementation is bad because sink_streams that
aren't ready to send start to recieve, even if there's
nothing to receive, and then they can't send because
they're waiting to receive.

This could be fixed a number of ways. If the sockets
lived permanently on the pollthread, and some identifier
was used to represent the socket on the futures runtime,
we could send messages requesting sends and receives
without one blocking the other.